Tpetra wrapped dual view microbenchmark (#9392)
* break out the bench from the unit tests
